create new file for meeting modal interfaces break all calendar components into their own file relocate large styles to their own stylesheet
