+++ src/interfaces/index.ts
@@ -1,5 +1,6 @@
-export * from './ProjectInterface'
-export * from './UserInterface'
-export * from './TicketInterFace'
-export * from './GenericInterfaces'
 export * from './CalendarInterfaces'
+export * from './GenericInterfaces'
+export * from './MeetingModalInterfaces'
+export * from './ProjectInterface'
+export * from './TicketInterFace'
+export * from './UserInterface'
