+++ src/screens/Calendar/MeetingModal/EditableMeetingModal/EditableMeetingModal.tsx
@@ -4,14 +4,9 @@ import {
   DialogContent,
   DialogActions,
   TextField,
-  FormControlLabel,
-  Checkbox,
-  Switch,
-  styled,
-  SwitchProps,
 } from '@mui/material'
 import { useEffect, useRef, useState } from 'react'
-import { BooleanObject, DateFieldsInterface } from 'interfaces'
+import { BooleanObject, DateFieldsInterface, EventInfo } from 'interfaces'
 import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
 import timezone from 'dayjs/plugin/timezone'
@@ -30,7 +25,6 @@ import {
   handleFormInputChange,
   initialDateFields,
 } from 'utils/helpers'
-import { AccessTime, Clear, People } from '@mui/icons-material'
 import { DescriptionField } from './MeetingTextField'
 import { initialMeetingText } from 'utils/data/calendarConstants'
 import {
@@ -40,30 +34,12 @@ import {
   setModalDisplayStatus,
   updateExistingEvent,
 } from 'utils/redux/slices/calendarSlice'
-import ArrowBackIcon from '@mui/icons-material/ArrowBack'
-import '../styles/MeetingModalStyles.scss'
-import { RiGoogleLine } from 'react-icons/ri'
+import '../styles/EditableMeetingModal.scss'
+import { MeetingModalHeaderIcons } from './MeetingModalHeaderIcons'
+import { GoogleMeetsToggler } from './GoogleMeetsToggler'
 
 dayjs.extend(utc)
 dayjs.extend(timezone)
-
-type AttendeeList = { email: string }
-interface EventInfo {
-  attendees: AttendeeList[]
-  description: string
-  googleMeetingInfo: {
-    enabled: boolean
-    hangoutLink?: string
-  }
-  end: {
-    dateTime: string
-  }
-  start: {
-    dateTime: string
-  }
-  summary: string
-  projectId: string
-}
 
 export const EditableMeetingModal = () => {
   const [meetingText, setMeetingText] = useState(initialMeetingText)
@@ -284,84 +260,6 @@ export const EditableMeetingModal = () => {
   )
 }
 
-const MeetingModalHeaderIcons = ({ handleClose }) => {
-  const dispatch = useAppDispatch()
-  const handleBackToDisplay = () => dispatch(setModalDisplayStatus('display'))
-
-  return (
-    <div className='meeting-modal-icons'>
-      <ArrowBackIcon
-        className='back-arrow-icon'
-        onClick={handleBackToDisplay}
-      />
-      <Clear className='clear-icon' onClick={handleClose} />
-    </div>
-  )
-}
-
-const GoogleMeetsToggler = ({ googleMeeting, toggleGoogleMeeting }) => {
-  const handleMeetToggler = () => {
-    toggleGoogleMeeting(!googleMeeting)
-  }
-
-  return (
-    <>
-      <RiGoogleLine className='google-meet-icon' size={23} />
-      <div className='google-meet-toggler'>
-        <FormControlLabel
-          control={
-            <IOSSwitch
-              sx={{ m: 1, marginLeft: '16px' }}
-              checked={googleMeeting}
-              onChange={handleMeetToggler}
-            />
-          }
-          label='Google Meet'
-          labelPlacement='start'
-          sx={{ margin: 0 }}
-        />
-        <p className='google-meet-text'>
-          A Google Meet link will be added to the calendar event.
-        </p>
-      </div>
-    </>
-  )
-}
-
-// Taken directly from MUI docs & modified to fit figma
-const IOSSwitch = styled((props: SwitchProps) => <Switch {...props} />)(() => ({
-  overflow: 'visible',
-  width: 32,
-  height: 16,
-  padding: 0,
-  '& .MuiSwitch-switchBase': {
-    padding: 0,
-    transitionDuration: '300ms',
-    '&.Mui-checked': {
-      transform: 'translateX(16px)',
-      color: '#fff',
-      '& + .MuiSwitch-track': {
-        backgroundColor: '#5C6BC0',
-        opacity: 1,
-        border: '0.5px solid #5C6BC0',
-      },
-    },
-  },
-  '& .MuiSwitch-thumb': {
-    boxShadow: 'none',
-    boxSizing: 'border-box',
-    border: '0.5px solid #5C6BC0',
-    marginTop: 0.5,
-    width: 16,
-    height: 16,
-  },
-  '& .MuiSwitch-track': {
-    border: '0.5px solid #5C6BC0',
-    borderRadius: 13,
-    backgroundColor: '#fff',
-    opacity: 1,
-  },
-}))
 const titleInputFieldStyles = {
   marginBottom: '32px',
   width: '100%',
