+++ src/screens/Calendar/MeetingModal/styles/DisplayMeetingModal.scss
