+++ src/screens/Calendar/MeetingModal/styles/EditableMeetingModal.scss
@@ -51,17 +51,12 @@
     .description-field {
       grid-area: description
     }
-
   }
-    .clock-icon {
-      grid-area: clockIcon;
-      margin-top: 10px;
-    }
 
-    .people-icon {
-      grid-area: peopleIcon;
-      margin-top: 10px;
-    }
+  .people-icon {
+    grid-area: peopleIcon;
+    margin-top: 10px;
+  }
 
   .meeting-modal-divider {
     background-color: #313335;
@@ -71,58 +66,12 @@
     right: 24px;
     width: 120%;
   }
-}
 
-/* Date Fields */
-.fields-wrapper {
-  display: flex;
-  flex-direction: column;
-  grid-area: dateFields;
-  justify-content:center;
-}
-
-.time-fields {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  height: 40px;
-
-  .time-icon {
-    position: relative;
-    right: 17px;
-    width: 17px;
+  .meeting-text-field {
+    display: flex;
+    flex-direction: row;
   }
-
-  span {
-    padding: 0 5px;
-  }
-  .span-dash {
-    font-size: large;
-    font-weight: 900;
-    background-color: 'black';
-    height: 15px;
-    width: 8px;
-  }
-}
-
-.timezone-wrapper {
-  align-self: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  width: 100%;
-   p {
-    font-size: 14px;
-    margin-right: 19px;
-   }
-}
-
-/* Meeting Text */ 
-.meeting-text-field {
-  display: flex;
-  flex-direction: row;
-}
-
+  
   .google-meet-icon {
     grid-area: meetIcon;
     margin-top: 5px;
@@ -137,3 +86,5 @@
   .google-meet-toggler {
     grid-area: meetToggler;
   }
+}
+
